move registry out of common to avoid cycle
