+++ pkg/registry/registry.go
@@ -1,4 +1,4 @@
-package common
+package registry
 
 import (
 	"context"
@@ -6,12 +6,11 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
-	"github.com/beam-cloud/beta9/pkg/metrics"
+	"github.com/beam-cloud/beta9/pkg/common"
 	"github.com/beam-cloud/beta9/pkg/types"
 	"github.com/google/uuid"
 	"github.com/rs/zerolog/log"
@@ -80,13 +79,16 @@ type ObjectStore interface {
 	Size(ctx context.Context, key string) (int64, error)
 }
 
+type S3Store struct {
+	client *s3.Client
+	config types.S3ImageRegistryConfig
+}
+
 func NewS3Store(config types.AppConfig) (*S3Store, error) {
-	cfg, err := GetAWSConfig(config.ImageService.Registries.S3.AccessKey, config.ImageService.Registries.S3.SecretKey, config.ImageService.Registries.S3.Region, config.ImageService.Registries.S3.Endpoint)
+	cfg, err := common.GetAWSConfig(config.ImageService.Registries.S3.AccessKey, config.ImageService.Registries.S3.SecretKey, config.ImageService.Registries.S3.Region, config.ImageService.Registries.S3.Endpoint)
 	if err != nil {
 		return nil, err
 	}
-
-	metricsRegistry := metrics.NewMetricsRegistry(config.Monitoring.VictoriaMetrics, "object_store")
 
 	return &S3Store{
 		client: s3.NewFromConfig(cfg, func(o *s3.Options) {
@@ -94,19 +96,11 @@ func NewS3Store(config types.AppConfig) (*S3Store, error) {
 				o.UsePathStyle = true
 			}
 		}),
-		config:          config.ImageService.Registries.S3,
-		metricsRegistry: metricsRegistry,
+		config: config.ImageService.Registries.S3,
 	}, nil
 }
 
-type S3Store struct {
-	client          *s3.Client
-	config          types.S3ImageRegistryConfig
-	metricsRegistry *metrics.MetricsRegistry
-}
-
 func (s *S3Store) Put(ctx context.Context, localPath string, key string) error {
-	start := time.Now()
 	f, err := os.Open(localPath)
 	if err != nil {
 		return err
@@ -123,16 +117,6 @@ func (s *S3Store) Put(ctx context.Context, localPath string, key string) error {
 		log.Error().Str("key", key).Err(err).Msg("error uploading image to registry")
 		return err
 	}
-
-	go func() {
-		info, err := os.Stat(localPath)
-		if err != nil {
-			log.Error().Str("key", key).Err(err).Msg("error getting file size")
-			return
-		}
-		sizeMB := float64(info.Size()) / 1024 / 1024
-		s.metricsRegistry.GetOrCreateHistogram("s3_upload_speed_mbps").Update(sizeMB / time.Since(start).Seconds())
-	}()
 
 	return nil
 }
